refactor: move reuse_prefix_cached_kv out of EC disagg mixin
Prefix-cache KV copy is a general prefill concern, unrelated to
encoder-cache disaggregation, so it does not belong in
ECDisaggHelpersMixin. Define reuse_prefix_cached_kv on the runner
itself; the EC consumer path still calls it via inheritance, and the
mixin only declares it as a host-provided method for type-checking.
